updated navbar
Hello friend your bootstrap project is very nice. But when I was checking though website, I found a mistake.
I am not able to redirect back to the page when I click on the welcome button on the navbar. So I corrected. If you like this then please merge my code